create banner component
