+++ src/components/organisms/Banner/Banner.css
@@ -1,0 +1,41 @@
+--bg-banner {
+  background-color: #f3f3f3;
+  margin-bottom: 40px;
+}
+.banner.--large {
+  width: 300px;
+  height: 600px;
+  background: #a0dbe9;
+  margin-bottom: 40px;
+}
+.banner.--top {
+  width: 320px;
+  height: 100px;
+  background: #a0dbe9;
+}
+.banner.--desktop {
+  display: none;
+}
+@media (min-width: 768px) {
+  .banner.--desktop {
+    display: none;
+  }
+  .banner.--top {
+    width: 728px;
+    height: 90px;
+  }
+}
+@media (min-width: 1024px) {
+  .banner.--desktop {
+    display: block;
+    margin: auto;
+  }
+  .banner.--top {
+    max-width: 1260px;
+    width: 100%;
+    height: 170px;
+  }
+}
+.sf-hidden {
+  display: none !important;
+}
