+++ src/components/templates/App/App.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+import Banner from "../../organisms/Banner/Banner";
+
+import "./App.css";
+
+function App() {
+  return (
+    <main className="App">
+      <Banner />
+    </main>
+  );
+}
+
+export default App;
