+++ src/index.css
@@ -1,13 +1,145 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+/* @media screen and (min-width: 20em) {
+  html {
+    font-size: calc(16px + (18 - 16) * ((100vw - 320px) / (1340 - 320)));
+  }
+}
+@media screen and (min-width: 22.5625em) {
+  html {
+    font-size: calc(16px + (18 - 16) * ((100vw - 360px) / (1340 - 360)));
+  }
+}
+@media screen and (min-width: 47.9375em) {
+  html {
+    font-size: calc(16px + (18 - 16) * ((100vw - 768px) / (1340 - 768)));
+  }
+}
+@media screen and (min-width: 64.0625em) {
+  html {
+    font-size: calc(16px + (18 - 16) * ((100vw - 1024px) / (1340 - 1024)));
+  }
+}
+@media screen and (min-width: 83.75em) {
+  html {
+    font-size: 16px;
+  }
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+.w-100 {
+  width: calc(100% + 20px) !important;
+  margin-left: -0.625em;
 }
+@media (min-width: 22em) {
+  .w-100 {
+    width: calc(100% + 40px) !important;
+    margin-left: -1.25em;
+  }
+}
+@media (min-width: 48em) {
+  .w-100 {
+    width: calc(100% + 60px) !important;
+    margin-left: -1.875em;
+  }
+}
+@media (min-width: 64em) {
+  .w-100 {
+    width: 100% !important;
+    margin-left: 0;
+  }
+}
+.w-100-mobile {
+  width: calc(100% + 20px) !important;
+  margin-left: -0.625em;
+}
+@media screen and (max-width: 47em) {
+  .w-100-mobile {
+    border: none !important;
+  }
+}
+@media (min-width: 22em) {
+  .w-100-mobile {
+    width: calc(100% + 40px) !important;
+    margin-left: -1.25em;
+  }
+}
+@media (min-width: 48em) {
+  .w-100-mobile {
+    width: 100% !important;
+    margin-left: 0;
+  }
+}
+@media (min-width: 64em) {
+  .w-100-mobile {
+    width: 100% !important;
+    margin-left: 0;
+  }
+}
+.row {
+  display: -ms-flexbox;
+  display: -webkit-box;
+  display: flex;
+  -ms-flex-wrap: wrap;
+  flex-wrap: wrap;
+  width: 100%;
+}
+
+.col-4,
+.col-12,
+.col-desksm-3,
+.col-desksm-4,
+.col-desksm-6,
+.col-desksm-9 {
+  position: relative;
+  width: 100%;
+}
+.col-4 {
+  -ms-flex: 0 0 33.33333%;
+  -webkit-box-flex: 0;
+  flex: 0 0 33.33333%;
+  max-width: 33.33333%;
+}
+.col-12 {
+  -ms-flex: 0 0 100%;
+  -webkit-box-flex: 0;
+  flex: 0 0 100%;
+  max-width: 100%;
+}
+@media (min-width: 64em) {
+  .col-desksm-3 {
+    -ms-flex: 0 0 25%;
+    -webkit-box-flex: 0;
+    flex: 0 0 25%;
+    max-width: 25%;
+  }
+  .col-desksm-4 {
+    -ms-flex: 0 0 33.33333%;
+    -webkit-box-flex: 0;
+    flex: 0 0 33.33333%;
+    max-width: 33.33333%;
+  }
+  .col-desksm-6 {
+    -ms-flex: 0 0 50%;
+    -webkit-box-flex: 0;
+    flex: 0 0 50%;
+    max-width: 50%;
+  }
+  .col-desksm-9 {
+    -ms-flex: 0 0 75%;
+    -webkit-box-flex: 0;
+    flex: 0 0 75%;
+    max-width: 75%;
+  }
+}
+@media (min-width: 80em) {
+  .col-deskxl-4 {
+    -ms-flex: 0 0 33.33333%;
+    -webkit-box-flex: 0;
+    flex: 0 0 33.33333%;
+    max-width: 33.33333%;
+  }
+  .col-deskxl-8 {
+    -ms-flex: 0 0 66.66667%;
+    -webkit-box-flex: 0;
+    flex: 0 0 66.66667%;
+    max-width: 66.66667%;
+  }
+} */
